quantum-edge-terminal/ai-engine/requirements.txt
fastapi==0.104.1
pydantic==2.5.0
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2
ta==0.10.2
aiohttp==3.9.1
asyncpg==0.29.0
redis==5.0.1
python-dotenv==1.0.0
psycopg2-binary==2.9.9
pyyaml==6.0.1
pytest==7.4.3
pytest-asyncio==0.21.1
